Nuså... ;)

+++ src/object/blinnphong.cpp
@@ -10,7 +10,7 @@ Vector3f BlinnPhong::brdf(const Vector3f &point, const Vector3f &dir, const Vect
 	Vector3f thetaN = dir.normalize();
 	Vector3f Hn = (Ln + thetaN).normalize();
 
-	double i = pow((Nn*Hn), roughness);
+	double i = powf((Nn*Hn), roughness);
 
 	Vector3f ret = getSpecular()*i*specColor*specIntensity + getDiffuse()*getColor();
